TravisCI integration (#1)
* Added rough first passes for Travis

Made up 2 test .travis.yml files to test when I have access to Heroku.

* Changed to Test Repo

Switched calls to Client facing API to test env.

* Added Basic Test

Created a trivial test case to see how/if everything is configured correctly.

* 1st Travis commit

After this commit, travis should be running.

* Attempting YML fix

There is a parsing error with the YML file. It might have something to do with tabs vs spaces.

* Switched to Heroku Deploy YML

Switched out the test YML file to use one of the Test Heroku Deploy scripts.

* Docker points to Dockerfile

The script wasn't pointing to the Dockerfile.

* Changed the wrong file.

Updated the file that is being used to have Docker point to the correct location.

* Added Heroku Login step

Not sure what Else will need to be done for Heroku Login.

* Heroku Login Attempt

Added some info back to deploy section. May need to configure Heroku somehow?

* Logged in via Heroku CLI

Logged into Heroku from the root project directory.

* Added GittHub connection to Heroku

Needed to connect GitHub repo with Travis associated to Heroku.

* Re-Encrypted API key

Could have been an issue where the settings are somehow encrypted in the key?

* Enabled option in Heroku

Enabled Automatic Deploys. Was connected to GitHub, but this option was not enabled.

* Still getting login Error

Followed some troubleshooting advice from here:
https://github.com/travis-ci/travis-ci/issues/6362

* Added Buildpack in Heroku Settings

Successfully got to the deploy stage. Heroku Errored since there was no selected Buildpack.

* Trying to go staright to Heroku

Will probably have to actually login, but attempting this anyway.

* Testing docker API Url

Soem documentation pages show heroku being updated through curl calls. Trying that.

* Trying to fix YML lint errors

* Fixed YML error

I hope.

* YML Linter says this is valid.

* Commented out offending? line

* Trying with Encrypted Env Var

Encrypted the HEROKU_API_KEY environmnet variable.

* Piping Secure Creds into login

Trying this, Travis hasn't built the last 2 commits...

* added stdin tag

Password should now be fed from command line

* Trying container login instead.

* Trying Cred Piping With no flag.

Looks like the --password-stdin flag was pointing to docker, not heroku. Trying to login with just piping.

* The Encrypted Env Car takes care of the password

Got a new error that indicates that you can't pipe in the password.

* The Env Var takes care of all Creds

Trying removing all Credential Piping.

* Trying OAuth method

This method uses temporary authorizations to allow access to the app for small windows of time. It sounds good, But I'm not hopeful...

* API Key Env Var may be blocking temp Auth

Got Errors about the HEROKU_API_KEY already being set. Using that for now, and if it works, I will check out the Temp Auth method.

* fixed some bash syntax

was using angled quotes, which indicates a command to be run, i think.

* Looks like Env Var will take care of password

Still getting Error about API Key being set. Might also be able to remove HEROKU_EMAIL.

* Removed Email Var and (re)Updated API Key

* Email Var required

Entered a waiting loop when no email provided. Now trying with just the updated key.

* Using Blank Emaill with API KEY

I think blank email may be a legacy requirement, but worth trying.

* Fixing YML Errors

* Actually fixed YML Error?

* Listing env vars on seperate lines is bad

Travis makes seperate jobs for every line of env vars. Fixed it so Email is getting set, and testing for which API key works

* Trying to figure out How multiple Env Vars work

* Fixed Key Encryption error

Forgot to encrypt the value as an Env Var

* Can't fun multiple tests at once?

* Got an Error from Run command

Run returned an error that it was unable to start worker processes, which may mean Auth was successful.

* Heroku isn't giving out worker threads?

Heroku uses something called Dyno's when calling the run command. It looks like either none are available, or they aren't there at all since this is the first deploy.

* Trying Temp Auth Method

Can't seem to get the API key to allow login. Trying with Temp Auth method.

* Documentation Doesn't Match

--output-format tag does not exist.

* Trying the OAuth Command again

This is the temp auth code. I don't think it will work...

* Invalid Syntax fix

I accidentally left a command in plain quotes.

* Trying Travis Auth again

Trying the (I think) last permutation of API_KEY and EMAIL variations. Checking what whoami returns in the travis environment.

* Error with API_KEY

* Trying with New Env Var

Checking to see if a Password var will work.

* Trying with correct Value for H_P var

Now using the auth:token as the password.

* Proving something to Ben

* Trying with Travis API key again

* Trying again

* Testing Login fudging

I Found a script that fudges the .netrc file that heroku writes to on login. Hopefully this allows me to go straight to heroku containers:login.

* Fixed Stray TAB

Tab character broke the Travis Build

* Actually fixed the tab.

* Added Docker Pushed to Heroku Registry

Docker Image needs to be tagged and pushed to Heroku before Heroku can push it the the web app. (As I understand it)

* Refactored to remove extra bash script

Moved the commands in docker-commands.sh into the script portion of of .travis.yml

* Pre-Build Move into Dir with Dockerfile

Might not even have to call 'docker build' since Heroku makes it sound like it does that for you.

* Added Release Call

heroku container:release is what actually pushes the image onto the app from the registry.

* Added Heroku CLI update

Need to update the Heroku CLI to make sure container:release is available.

* Updated README and cleaned up root

Moved all old/testing .travis.yml files into the travis tests folder. Cleaned up the final .travis.yml file. Updated the README to contain the current steps.

* Published 214 and 215 branches

Deleted 214 and 215 folders. They are in their own branch now.

* Heroku Rejects the Deploy

Heroku asks for its own manifest or that the image builds on the same stack as the app. I chose to switch stacks since the manifest appears to expect a CMD in the Dockerfile.

* Added Missing Tag When setting Stack

Was missing application tag on the stack:set command.

* Build Error Message Still Exists

On Heroku, it says it is deploying, but it also reports a build failure with the same timestamp.

* Updated README

The Heroku Build issue was a setting in Heroku. It was trying to build and deploy on its own, but the Travis script was already deploying the App.

* Fixed README markdown

* Removed heroku.yml

Don't need this file. It was added due to the Heroku Setting Error referenced in the past few commits.

* Removed travis-tests and Cleaned up comments

* Added Email Notification Settings

Using the email tag will send an email to everyone listed. I have it set to send an email on every success and every fail.

* Refactored so all Files are in root

Removed Main file. Now all needed files are in the root directory.

* Cleaned up Travis file

Removed calls that were not needed and cleaned up comments.